--- JDesignPatterns/src/dp/creational/BigMac.java
@@ -1,11 +1,0 @@
-
-package dp.creational;
-
-public class BigMac extends Sanduiche{
-    
-    public BigMac(){
-        nome = "BigMac";
-        pao = "com gergelin";
-        carne = "100g dupla";
-    }    
-}
