--- JDesignPatterns/src/dp/creational/BigTasty.java
@@ -1,12 +1,0 @@
-
-package dp.creational;
-
-public class BigTasty extends Sanduiche{
-    
-    public BigTasty(){
-        nome = "BigTasty";
-        pao = "com gergelin";
-        carne = "de 200g";
-    }   
-    
-}
