--- JDesignPatterns/src/dp/creational/BurguerKing.java
@@ -1,14 +1,0 @@
-
-package dp.creational;
-
-public class BurguerKing extends HamburguerStore{
-    @Override
-    Sanduiche createHamburguer(String item){
-        switch (item) {
-            case "Whopper":
-                return new Whopper();
-            default:
-                return null;   
-        }
-    } 
-}
