--- JDesignPatterns/src/dp/creational/CheddarMcMelt.java
@@ -1,13 +1,0 @@
-
-package dp.creational;
-
-public class CheddarMcMelt extends Sanduiche{
-    
-    public CheddarMcMelt(){
-        nome = "CheddarMcMelt";
-        pao = "com gergelin";
-        carne = "100g única";
-    }   
-    
-    
-}
