--- JDesignPatterns/src/dp/creational/FactoryMethod.java
@@ -1,13 +1,0 @@
-
-package dp.creational;
-
-
-public class FactoryMethod {
-
-    public static void main(String[] args) {
-        
-        HamburguerStore nyMcStore = new McDonalds();
-        nyMcStore.orderHamburguer("BigMac");
-        
-    }
-}
