--- JDesignPatterns/src/dp/creational/HamburguerStore.java
@@ -1,29 +1,0 @@
-
-package dp.creational;
-
-public abstract class HamburguerStore{
-    
-    Sanduiche createHamburguer(String type){        
-        return null;
-    }
-    
-    public Sanduiche orderHamburguer(String type){
-        
-        Sanduiche hamburguer;
-        
-        hamburguer = createHamburguer(type);
-        
-        hamburguer.prepare();
-        hamburguer.assar();
-        hamburguer.fritar();
-        hamburguer.embalar();
-        
-        
-        return hamburguer;
-        
-    }
-    
-    
-    
-}
-
