--- JDesignPatterns/src/dp/creational/McDonalds.java
@@ -1,21 +1,0 @@
-
-package dp.creational;
-
-public class McDonalds extends HamburguerStore {
-    
-    @Override
-    Sanduiche createHamburguer(String item){
-        switch (item) {
-            case "BigMac":
-                return new BigMac();
-            case "Quarteirao":
-                return new Quarteirao();
-            case "Big Tasty":
-                return new BigTasty();
-            case "Cheddar McMelt":
-                return new CheddarMcMelt();
-            default:
-                return null;   
-        }
-    } 
-}
