--- JDesignPatterns/src/dp/creational/Quarteirao.java
@@ -1,12 +1,0 @@
-
-package dp.creational;
-
-public class Quarteirao extends Sanduiche{
-    
-    public Quarteirao(){
-        nome = "Quarteirão";
-        pao = "com gergelin";
-        carne = "de 180g";
-    }   
-    
-}
