--- JDesignPatterns/src/dp/creational/Whopper.java
@@ -1,10 +1,0 @@
-package dp.creational;
-
-public class Whopper extends Sanduiche{
-    
-    public Whopper(){
-        nome = "Whopper®";
-        pao = "com gergelin";
-        carne = "100g dupla";
-    }   
-}
