+++ JDesignPatterns/src/dp/creational/factorymethod/BigMac.java
@@ -1,0 +1,11 @@
+
+package dp.creational.factorymethod;
+
+public class BigMac extends Sanduiche{
+    
+    public BigMac(){
+        nome = "BigMac";
+        pao = "com gergelin";
+        carne = "100g dupla";
+    }    
+}
