+++ JDesignPatterns/src/dp/creational/factorymethod/BigTasty.java
@@ -1,0 +1,12 @@
+
+package dp.creational.factorymethod;
+
+public class BigTasty extends Sanduiche{
+    
+    public BigTasty(){
+        nome = "BigTasty";
+        pao = "com gergelin";
+        carne = "de 200g";
+    }   
+    
+}
