+++ JDesignPatterns/src/dp/creational/factorymethod/CheddarMcMelt.java
@@ -1,0 +1,13 @@
+
+package dp.creational.factorymethod;
+
+public class CheddarMcMelt extends Sanduiche{
+    
+    public CheddarMcMelt(){
+        nome = "CheddarMcMelt";
+        pao = "com gergelin";
+        carne = "100g única";
+    }   
+    
+    
+}
