+++ JDesignPatterns/src/dp/creational/factorymethod/FactoryMethod.java
@@ -1,0 +1,13 @@
+
+package dp.creational.factorymethod;
+
+
+public class FactoryMethod {
+
+    public static void main(String[] args) {
+        
+        HamburguerStore nyMcStore = new McDonalds();
+        nyMcStore.orderHamburguer("BigMac");
+        
+    }
+}
