+++ JDesignPatterns/src/dp/creational/factorymethod/HamburguerStore.java
@@ -1,0 +1,29 @@
+
+package dp.creational.factorymethod;
+
+public abstract class HamburguerStore{
+    
+    Sanduiche createHamburguer(String type){        
+        return null;
+    }
+    
+    public Sanduiche orderHamburguer(String type){
+        
+        Sanduiche hamburguer;
+        
+        hamburguer = createHamburguer(type);
+        
+        hamburguer.prepare();
+        hamburguer.assar();
+        hamburguer.fritar();
+        hamburguer.embalar();
+        
+        
+        return hamburguer;
+        
+    }
+    
+    
+    
+}
+
