+++ JDesignPatterns/src/dp/creational/factorymethod/Quarteirao.java
@@ -1,0 +1,12 @@
+
+package dp.creational.factorymethod;
+
+public class Quarteirao extends Sanduiche{
+    
+    public Quarteirao(){
+        nome = "Quarteirão";
+        pao = "com gergelin";
+        carne = "de 180g";
+    }   
+    
+}
