+++ JDesignPatterns/src/dp/creational/factorymethod/Whopper.java
@@ -1,0 +1,10 @@
+package dp.creational.factorymethod;
+
+public class Whopper extends Sanduiche{
+    
+    public Whopper(){
+        nome = "Whopper®";
+        pao = "com gergelin";
+        carne = "100g dupla";
+    }   
+}
